# Backend/openapi.yaml
openapi: 3.0.0
info:
  title: ML Model Serving Platform
  description: |
    # ML Model Serving Platform API

    A production-ready platform for deploying, managing, and serving machine learning models via REST API.

    ## Key Features

    - **Model Management**: Upload, version, and manage ML models
    - **Authentication**: JWT + API Key authentication with refresh tokens
    - **Predictions**: Real-time predictions with caching and performance tracking
    - **Sharing**: Share models with other users with granular permissions
    - **Analytics**: Track model performance and usage metrics
    - **Webhooks**: Event-driven architecture with webhook support
    - **Rate Limiting**: Per-user rate limiting with Redis backend
    - **Health Checks**: Kubernetes-ready health check endpoints

    ## Getting Started

    1. **Register**: `POST /api/v1/auth/register`
    2. **Login**: `POST /api/v1/auth/login`
    3. **Upload Model**: `POST /api/v1/models/upload`
    4. **Make Prediction**: `POST /api/v1/predictions/predict`

  version: "1.0.0"
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
servers:
  - url: "https://api.example.com"
    description: Production server
  - url: "http://localhost:8000"
    description: Local development server
...
paths:
  /api/v1/auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  example: secure_password
                full_name:
                  type: string
                  example: John Doe
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  email:
                    type: string
                  full_name:
                    type: string
                  created_at:
                    type: string
                    format: date-time
  /api/v1/auth/login:
    post:
      summary: Login user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  token_type:
                    type: string
                    example: bearer
                  expires_in:
                    type: integer
  /api/v1/auth/refresh:
    post:
      summary: Refresh access token
      tags:
        - Authentication
      security:
        - BearerToken: []
      responses:
        "200":
          description: New access token generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                  expires_in:
                    type: integer
  /api/v1/auth/me:
    get:
      summary: Get current user
      tags:
        - Authentication
      security:
        - BearerToken: []
      responses:
        "200":
          description: Current user details
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  email:
                    type: string
                  full_name:
                    type: string
                  created_at:
                    type: string
                    format: date-time
...
  /api/v1/models:
    get:
      summary: List user's models
      tags:
        - Models
      security:
        - BearerToken: []
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          schema:
            type: string
            enum:
              - active
              - deprecated
              - archived
      responses:
        "200":
          description: List of models
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
                  skip:
                    type: integer
                  limit:
                    type: integer
...
  /api/v1/model-shares/shared-with-me:
    get:
      summary: List models shared with me
      tags:
        - Model Sharing
      security:
        - BearerToken: []
      responses:
        "200":
          description: List of shared models
...
components:
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login

    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key from /api-keys/create
...